+++ admin/controller/extension/payment/mollie_voucher.php
@@ -1,0 +1,4 @@
+<?php
+include_once(DIR_APPLICATION . "controller/payment/mollie_voucher.php");
+	class ControllerExtensionPaymentMollieVoucher extends ControllerPaymentMollieVoucher{}
+?>