+++ admin/language/english/payment/mollie.php
@@ -46,9 +46,12 @@ $_['text_mollie_giftcard']      = $method_list_logo;
 $_['text_mollie_eps']           = $method_list_logo;
 $_['text_mollie_giropay']       = $method_list_logo;
 $_['text_mollie_klarnapaylater'] = $method_list_logo;
+$_['text_mollie_klarnapaynow']   = $method_list_logo;
 $_['text_mollie_klarnasliceit']  = $method_list_logo;
 $_['text_mollie_przelewy24']  	 = $method_list_logo;
 $_['text_mollie_applepay']  	 = $method_list_logo;
+$_['text_mollie_voucher']    	 = $method_list_logo;
+
 //These methods have been removed, it's here only to supress error of undefined variables. We'll remove it later
 $_['text_mollie_inghomepay']  	 = $method_list_logo;
 $_['text_mollie_directdebit']  	 = $method_list_logo;
@@ -75,9 +78,12 @@ $_['name_mollie_giftcard']      = 'Giftcard';
 $_['name_mollie_eps']           = 'EPS';
 $_['name_mollie_giropay']       = 'Giropay';
 $_['name_mollie_klarnapaylater'] = 'Klarna Pay Later';
+$_['name_mollie_klarnapaynow']   = 'Klarna Pay Now';
 $_['name_mollie_klarnasliceit']  = 'Klarna Slice It';
 $_['name_mollie_przelewy24']  	 = 'P24';
 $_['name_mollie_applepay']  	 = 'Apple Pay';
+$_['name_mollie_voucher']        = "Voucher";
+
 //These methods have been removed, it's here only to supress error of undefined variables. We'll remove it later
 $_['name_mollie_inghomepay']  	 = 'INGHOMEPAY';
 $_['name_mollie_directdebit']  	 = 'Direct Debit';
@@ -124,6 +130,9 @@ $_['text_more']                    = 'More';
 $_['text_no_maximum_limit']        = 'No maximum amount limit';
 $_['text_standard_total']          = 'Standard Total: %s';
 $_['text_advance_option']          = 'Advanced Options For %s';
+$_['text_payment_api']             = 'Payments API';
+$_['text_order_api']               = 'Orders API';
+$_['text_info_orders_api']         = 'Why Use Orders API?';
 
 // Entry
 $_['entry_payment_method']           = "Payment method";
@@ -165,6 +174,7 @@ $_['entry_payment_fee_tax_class']    = "Payment Fee Tax Class";
 $_['entry_total']				     = "Total";
 $_['entry_minimum']				     = "Minimum";
 $_['entry_maximum']				     = "Maximum";
+$_['entry_api_to_use']  		     = "API to Use";
 
 // Help
 $_['help_view_profile']             = 'You can find your API key in <a href="https://www.mollie.com/dashboard/settings/profiles/" target="_blank" class="alert-link">your Mollie website profiles</a>.';
@@ -228,6 +238,7 @@ $_['entry_create_shipment_on_order_complete'] 		= "Create shipment upon order co
 $_['button_update'] = "Update";
 $_['button_mollie_connect'] = "Connect via Mollie";
 $_['button_advance_option'] = "Advance Option";
+$_['button_close'] = "Close";
 
 //Error log
 $_['text_log_success']	   = 'Success: You have successfully cleared your mollie log!';