+++ admin/language/french/payment/mollie.php
@@ -46,9 +46,12 @@ $_['text_mollie_giftcard']      = $method_list_logo;
 $_['text_mollie_eps']           = $method_list_logo;
 $_['text_mollie_giropay']       = $method_list_logo;
 $_['text_mollie_klarnapaylater'] = $method_list_logo;
+$_['text_mollie_klarnapaynow']   = $method_list_logo;
 $_['text_mollie_klarnasliceit']  = $method_list_logo;
 $_['text_mollie_przelewy24']  	 = $method_list_logo;
 $_['text_mollie_applepay']  	 = $method_list_logo;
+$_['text_mollie_voucher']    	 = $method_list_logo;
+
 //These methods have been removed, it's here only to supress error of undefined variables. We'll remove it later
 $_['text_mollie_inghomepay']  	 = $method_list_logo;
 $_['text_mollie_directdebit']  	 = $method_list_logo;
@@ -74,10 +77,12 @@ $_['name_mollie_sofort']        = "SOFORT Banking";
 $_['name_mollie_giftcard']      = 'Giftcard';
 $_['name_mollie_eps']           = 'EPS';
 $_['name_mollie_giropay']       = 'Giropay';
+$_['name_mollie_klarnapaynow']   = 'Klarna Pay Now';
 $_['name_mollie_klarnapaylater'] = 'Klarna Pay Later';
 $_['name_mollie_klarnasliceit']  = 'Klarna Slice It';
 $_['name_mollie_przelewy24']  	 = 'P24';
 $_['name_mollie_applepay']  	 = 'Apple Pay';
+$_['name_mollie_voucher']        = "Voucher";
 //These methods have been removed, it's here only to supress error of undefined variables. We'll remove it later
 $_['name_mollie_inghomepay']  	 = 'INGHOMEPAY';
 $_['name_mollie_directdebit']  	 = 'Direct Debit';
@@ -124,6 +129,9 @@ $_['text_more']                    = 'Suite';
 $_['text_no_maximum_limit']        = 'Pas de limite de montant maximum';
 $_['text_standard_total']          = 'Total standard: %s';
 $_['text_advance_option']          = 'Options avancées pour %s';
+$_['text_payment_api']             = 'API de paiement';
+$_['text_order_api']               = 'API de commandes';
+$_['text_info_orders_api']         = 'Pourquoi utiliser l\'API des commandes?';
 
 // Entry
 $_['entry_payment_method']           = "Procédé de paiement";
@@ -164,6 +172,7 @@ $_['entry_payment_fee_tax_class']    = "Payment Fee Tax Class";
 $_['entry_total']				     = "Le total";
 $_['entry_minimum']				     = "Le minimum";
 $_['entry_maximum']				     = "Maximum";
+$_['entry_api_to_use']  		     = "API à utiliser";
 
 // Help
 $_['help_view_profile']             = 'Vous pouvez trouver votre clé API dans <a href="https://www.mollie.com/dashboard/settings/profiles/" target="_blank" class="alert-link">vos profils de site à Mollie</a>.';
@@ -227,6 +236,7 @@ $_['entry_create_shipment_on_order_complete'] 		= "Créer une expédition à la 
 $_['button_update']         = "Mettre à jour";
 $_['button_mollie_connect'] = "Connect via Mollie";
 $_['button_advance_option'] = "Advance Option";
+$_['button_close'] = "Fermer";
 
 //Error log
 $_['text_log_success']	   = 'Succès: vous avez effacé avec succès votre journal des erreurs!';