+++ catalog/controller/extension/payment/mollie_klarnapaynow.php
@@ -1,0 +1,5 @@
+<?php
+include_once(DIR_APPLICATION . "controller/payment/mollie_klarnapaynow.php");
+class ControllerExtensionPaymentMollieKlarnapaynow extends ControllerPaymentMollieKlarnapaynow
+{
+}