+++ catalog/model/extension/payment/mollie_voucher.php
@@ -1,0 +1,5 @@
+<?php
+include_once(DIR_APPLICATION . "model/payment/mollie_voucher.php");
+class ModelExtensionPaymentMollieVoucher extends ModelPaymentMollieVoucher
+{
+}